added cores that were labeled in opposite order
These .csv files have 4 columns in the data.   Column "actual" is what
each row should be labeled for ring width calculations.  I kept the
other column because that is what the points on the image show as.  I
will delete this subfolder and delete the unwanted column after resin
duct numbers are counted for each tree ring.  Once that is done, I will
add the updated .csv file to the ring_coordinates folder.