Created radius column in rings.R.  Added notes
Created the radius column to reflect radius for each tree ring.
ring.dist recorded distance between two points, but did not reflect the
actual radius for the corresponding ring.  Created shift function to
accomplish this.  Also, added notes for work needed to accomplish in the
code to obtain ring area data.  Will need to talk to Dr. Schwilk about
this.