Writing script for ring.area
Beginning script for ring.area.  I am currently trying to run it in a
loop, but I'm encountering issues.  There also might be a problem with
the core.widths function that I wrote.  It seems like that is where the
error is coming from when I run the loop.  There is definitely an easier
way to do this without as much coding, but it's a start.