Added precipitation weather stations
Added more stations that were located within each city that the mountain
ranges were located in.  Code goes through and determines the station
that is closest to the tagged tree and assigns the yearly value.  I
updated the .csv file located on dropbox as well to reflect all of the
new values.