Add # months for each precip year
I noticed that when I downloaded annual precip data, some of the years
contained less than 12 months worth of data.  The number of months used
to calculate the yearly precipitation amount for an area is now
indicated in a new column.  A decision regarding what is an acceptable #
of months of data needs to be established now, because some stations
have less than 6 months of data for a yearly amount.