Add csv file for palmer drought values
Obtained values for palmer drought severity index, as well as other data
from NOAA for the trans pecos area.  This data is not specific to
mountain range, but it was the best I was able to find.  Specific
mountain range questions can't be answered with this, but an overall
effect can be looked at with the values.  Will add a metadata csv file
soon.