Added drought values
Added regional drought values associated with the Trans Pecos region to
each tree ring that had a calendar year that matched drought year
values.  Also changed the plot at the end to go by ring instead of age
because age refers to the actual age of the ring, and ring refers to
tree age.