Script modifications
Modified two scripts.  Read_rings.R calculated precipitation values from
multiple precip stations.  It was determined that this will be
ineffective because there isn't enough data, and the data doesn't seem
to follow exact patterns in the mountain ranges, so I commented the
lines of code.  I chose to still leave them in the script in case we
determine that the data obtained from precip stations may be relevent
later on.

In read_precip_drought_data I only changed a few lines with the comments
in the beginning and the ggplot that was produced at the end.