Created new column with 3 yr lag for drought
Modified script in read_precip_drought_data to include a newly
calculated column called PMDI_3yrlag which takes the value of the
current year, plus the two previous year PMDI values and calculates the
mean.  Doing so removes the two earliest data years because, but only
those two are removed.  We can decide later on if we believe this is the
best approach to go about assigning PMDI values.