Added line so code will run with new column
Can't mutate non-numeric values, so since I added a new column into
rings.R, it needed to be reflected in the script that called it.  I will
be editing this script and pulling some graphs from it into lmer mods,
but I am going to keep it open for the time being.