Added mdata dataframe creation to read_rings
All data analyses ran code from read_rings.R to create an "mdata"
dataframe.  Decided to have this in read_rings to have that r script be
the only one creating new dataframes.  Also manually converted the
scaled values that were created from matrix to numeric values.  This was
done very inefficiently, but after spending a lot of time trying to find
a quick solution to do it, I gave up in interest of time so I could
graph out some of the variables.  Dplyr will not work with columns that
are in class: matrix inside of a dataframe which is why I did it
manually.  Will open up an issue to get that fixed later.