Removed mdata creation and added analyses
Removed code to create mdata dataframe since that is now being created
in read_rings.R.  Also removed graphs from this R script because this is
going to be the script that only has analyses, and a seperate script
will contain graphs.  Some decisions need to be made regarding which
models are the best fit for the data and the story we are trying to
tell.  Will open up a github issue to discuss this.