Update lmer models with new models
Removed the old models that looked at individual species and used BAI.
Added 4 new models at the end that are run with a subset of the data
(ponderosae trees and cembroides trees).  I've indicated in the comments
the signficant effects that the model suggests, but I don't know if they
are running correctly.  I've indicated some of the error messages that i
get when I try running them, so I don't know if the results are truly
indicative of what I am trying to test.

Also, I added species as a random effect in the first two models that
looked at subsections.  AIC values gave no indication that adding
species as a random effect produces a better model, so I don't know if
it is best to include that or not.