Fixed full and simple models
Fixed the model specifications to include the overall effects even when
only the interaction effects were deemed signficant.

I had aimed to do this in an earlier commit, but I accidently commited a
bunch of files that I do not want under version control because they
were me playing with data and creating temporary graphs.  That's why
there's a bunch of revert commits before this.