Create new data frame with coefficients from summary
Created a new dataframe that contained the coefficients from summary of
the linear model.  Also created a new column calculating the pvalue
based on the t values given from the summary.   I don't know if it is
the best calculation though, because it assumes normal distribution
which this dataset deviates from a bit.  If there is a more conservative
method to calcualting p value, I'd probably like to use that one.