Update lmer models
Lmer models were updated with post-hoc testing on interaction terms.
Post-hoc testing cannot be done on three way interactions, though, so
those relationships will just have to be demonstrated with pvalues and
graphical relationships.

Cleaned the script up a little bit to run better.  Will be doing more to
it later to make it look nicer.